1 Translate the fr.elimerl.registre.model package.
